+++ IspDataDb/dbo/Tables/Country.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[Country] (
-    [IDNo]           INT            IDENTITY (1, 1) NOT NULL,
+    [IDNo]           SMALLINT       IDENTITY (1, 1) NOT NULL,
     [ISOA2]          VARCHAR (2)    CONSTRAINT [DF__Countries__count__2A4B4B5E] DEFAULT ('') NOT NULL,
     [CountryName]    VARCHAR (100)  CONSTRAINT [DF__Countries__count__2B3F6F97] DEFAULT ('') NOT NULL,
     [CountryNameAra] NVARCHAR (200) CONSTRAINT [DF__Countries__count__2C3393D0] DEFAULT ('') NOT NULL,
@@ -8,7 +8,7 @@
     [Flag32]         VARCHAR (256)  NULL,
     [Flag128]        VARCHAR (256)  NULL,
     [ISOA3]          VARCHAR (3)    NULL,
-    [ISON]           INT            NULL,
+    [ISON]           SMALLINT       NULL,
     [PhoneCode]      VARCHAR (4)    NULL,
     [DateTimeStamp]  ROWVERSION     NULL,
     CONSTRAINT [PK_CountryIDNo] PRIMARY KEY CLUSTERED ([IDNo] ASC),
@@ -16,4 +16,6 @@
     CONSTRAINT [IX_NameAra] UNIQUE NONCLUSTERED ([IDNo] ASC),
     CONSTRAINT [IX_NameEng] UNIQUE NONCLUSTERED ([CountryName] ASC)
 );
+
+
