+++ IspDataDb/dbo/Tables/DistributionSchemeItem.sql
@@ -1,9 +1,11 @@
 ﻿CREATE TABLE [dbo].[DistributionSchemeItem] (
     [IdNo]                   INT             IDENTITY (1, 1) NOT NULL,
     [DistributionSchemeIdNo] INT             NULL,
-    [Sequence]               INT             NULL,
-    [RevCostCenterIdNo]       INT             NULL,
+    [Sequence]               SMALLINT        NULL,
+    [RevCostCenterIdNo]      SMALLINT        NULL,
     [Percentage]             DECIMAL (18, 2) NULL,
     CONSTRAINT [PK_DistributionSchemeItem] PRIMARY KEY CLUSTERED ([IdNo] ASC)
 );
+
+
