+++ IspDataDb/dbo/Tables/EmployeeAbsencesLeave.sql
@@ -1,9 +1,11 @@
 ﻿CREATE TABLE [dbo].[EmployeeAbsencesLeave] (
     [IdNo]             INT      NULL,
-    [EmployeeIdNo]     INT      NULL,
+    [EmployeeIdNo]     SMALLINT NULL,
     [LeaveAbsenceIdNo] SMALLINT NULL,
     [StartDate]        DATE     NULL,
     [EndDate]          DATE     NULL,
     [Approved]         BIT      NULL
 );
+
+
