+++ IspDataDb/dbo/Tables/ErJournal.sql
@@ -1,11 +1,11 @@
 ﻿CREATE TABLE [dbo].[ErJournal] (
     [IDNo]            INT            IDENTITY (1, 1) NOT NULL,
-    [EmployeeIdNo]    INT            NOT NULL,
+    [EmployeeIdNo]    SMALLINT       NOT NULL,
     [TransactionDate] DATE           NULL,
     [ReferenceNo]     VARCHAR (15)   NULL,
     [TransactionType] CHAR (1)       NULL,
     [Amount]          MONEY          NOT NULL,
-    [AccountIdNo]     INT            NOT NULL,
+    [AccountIdNo]     SMALLINT       NOT NULL,
     [Notes]           NVARCHAR (255) NOT NULL,
     [Posted]          BIT            NULL,
     [Cancelled]       BIT            NULL,
@@ -13,6 +13,8 @@
     [DateTimeStamp]   ROWVERSION     NOT NULL,
     CONSTRAINT [PK_ErIdNo] PRIMARY KEY CLUSTERED ([IDNo] ASC)
 );
+
+
