+++ IspDataDb/dbo/Tables/ReceiptDetails.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[ReceiptDetails] (
-    [IdNo]               BIGINT       IDENTITY (1, 1) NOT NULL,
-    [AccountIdNo]        INT          NULL,
+    [IdNo]               INT          IDENTITY (1, 1) NOT NULL,
+    [AccountIdNo]        SMALLINT     NULL,
     [CheckPayment]       BIT          NULL,
     [CheckReferenceNo]   VARCHAR (15) COLLATE SQL_Latin1_General_CP1_CI_AS NULL,
     [ORNumber]           VARCHAR (15) COLLATE SQL_Latin1_General_CP1_CI_AS NULL,
@@ -10,4 +10,6 @@
     [UnApplied]          MONEY        NULL,
     CONSTRAINT [PK_ReceiptDetails] PRIMARY KEY CLUSTERED ([IdNo] ASC)
 );
+
+
