+++ IspDataDb/dbo/Tables/SJPrefix.sql
@@ -1,7 +1,9 @@
 ﻿CREATE TABLE [dbo].[SJPrefix] (
     [IdNo]        TINYINT  NOT NULL,
-    [AccountIdNo] INT      NULL,
+    [AccountIdNo] SMALLINT NULL,
     [Prefix]      CHAR (1) COLLATE SQL_Latin1_General_CP1_CI_AS NULL,
     CONSTRAINT [PK_SJPrefix] PRIMARY KEY CLUSTERED ([IdNo] ASC)
 );
+
+
