+++ IspDataDb/dbo/Tables/User.sql
@@ -1,8 +1,8 @@
 ﻿CREATE TABLE [dbo].[User] (
-    [IdNo]              INT           IDENTITY (18, 1) NOT NULL,
+    [IdNo]              SMALLINT      IDENTITY (18, 1) NOT NULL,
     [UserName]          VARCHAR (20)  NOT NULL,
     [Password]          VARCHAR (50)  NULL,
-    [SecurityGroupIDNo] INT           NULL,
+    [SecurityGroupIDNo] SMALLINT      NULL,
     [FullName]          VARCHAR (50)  NULL,
     [FullNameAra]       NVARCHAR (50) NULL,
     [SecurityLevel]     TINYINT       NULL,
@@ -10,4 +10,6 @@
     CONSTRAINT [PK_UserIDNo] PRIMARY KEY CLUSTERED ([IdNo] ASC),
     CONSTRAINT [FK_User_User] FOREIGN KEY ([IdNo]) REFERENCES [dbo].[User] ([IdNo])
 );
+
+
